% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_boot_mean.R
\name{cl_boot_mean}
\alias{cl_boot_mean}
\title{CL boot mean}
\usage{
cl_boot_mean(times, x, id, pval)
}
\arguments{
\item{times}{Numero de drawns}

\item{x}{variavel de interesse}

\item{id}{cluster}

\item{pval}{CI}
}
\description{
Funcao para calcular CI de medias com cluster bootstrap
}
