% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg_felm.R
\name{marg_felm}
\alias{marg_felm}
\title{Marginal Effects para felm}
\usage{
marg_felm(model, base_name, rem_name, ci = 0.95)
}
\arguments{
\item{model}{Um modelo estimado com o pacote lfe}

\item{base_name}{Um character com o nome do coeficiente do tratamento}

\item{rem_name}{Um character com o nome da parte dos coeficientes para remover}

\item{ci}{Intervalo de confianca}
}
\description{
Funcao para retornar efeitos marginal para modelos felm com interacoes entre tratamento e variavel discreta
}
