% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bw.R
\name{plot_bw}
\alias{plot_bw}
\title{BW plot}
\usage{
plot_bw(y, x, c = 0, p = 1, h_bw = NULL, cluster = NULL, level = 95)
}
\arguments{
\item{y}{Var dependente;}

\item{x}{scores;}

\item{c}{Cutpoint;}

\item{p}{Polinomio (default 1);}

\item{h_bw}{Vetor com os valores inicial, final e intervalos do BW;}

\item{cluster}{Cluster robust SE;}

\item{level}{CI;}
}
\value{
A ggplot object;
}
\description{
Funcao para plotar sensibilidade do estimador RD
}
