% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_point.R
\name{point_rd}
\alias{point_rd}
\title{Point estimate RD}
\usage{
point_rd(modelos, vars)
}
\arguments{
\item{modelos}{Lista com modelos estimados com rdrobust}

\item{vars}{Vetor com o nome das variaveis}
}
\description{
Plota varios estimadores RD lado a lado
}
