% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd_tables.R
\name{rd_horizontal}
\alias{rd_horizontal}
\title{RD horizontal}
\usage{
rd_horizontal(modelos, vars)
}
\arguments{
\item{modelos}{Uma lista contendo os modelos}

\item{vars}{Um vetor contendo o nome das variaveis}
}
\description{
Funcao para resumir varios modelos RD em linhas horizontais
}
