% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd_tables.R
\name{rd_internal}
\alias{rd_internal}
\title{RD interna}
\usage{
rd_internal(modelo, var, vert = T)
}
\arguments{
\item{modelo}{Um modelo estimado com o pacote rdrobust}

\item{var}{Um character com o nome do modelo}

\item{vert}{O modelo pode ser reportado na vertical?}
}
\description{
Funcao para resumir modelos RD (com notacao para LaTeX)
}
