% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdd_tables.R
\name{rd_vertical}
\alias{rd_vertical}
\title{RD vertical}
\usage{
rd_vertical(modelos, vars)
}
\arguments{
\item{modelos}{Uma lista contendo os modelos}

\item{vars}{Um vetor contendo o nome das variaveis}
}
\description{
Funcao para resumir varios modelos RD em colunas verticais
}
